+++ BriansUsbQuizBoxApi/GameDoneEventArgs.cs
@@ -8,9 +8,49 @@ namespace BriansUsbQuizBoxApi
     public class GameDoneEventArgs : EventArgs
     {
         /// <summary>
-        /// Winner paddle that was pressed. Null if no winner
+        /// First place winner paddle that was pressed. Null if no winner
         /// </summary>
         public Paddle? Winner { get; private set; }
+
+        /// <summary>
+        /// True if additional winner information (beyond first place winner) is included, otherwise false
+        /// </summary>
+        public bool HasAdditionalWinnerInfo { get; private set; }
+
+        /// <summary>
+        /// Second place winner paddle that was pressed. Null if no winner
+        /// </summary>
+        public Paddle? Winner2 { get; private set; }
+
+        /// <summary>
+        /// Third place winner paddle that was pressed. Null if no winner
+        /// </summary>
+        public Paddle? Winner3 { get; private set; }
+
+        /// <summary>
+        /// Fourth place winner paddle that was pressed. Null if no winner
+        /// </summary>
+        public Paddle? Winner4 { get; private set; }
+
+        /// <summary>
+        /// Fifth place winner paddle that was pressed. Null if no winner
+        /// </summary>
+        public Paddle? Winner5 { get; private set; }
+
+        /// <summary>
+        /// Sixth place winner paddle that was pressed. Null if no winner
+        /// </summary>
+        public Paddle? Winner6 { get; private set; }
+
+        /// <summary>
+        /// Seventh place winner paddle that was pressed. Null if no winner
+        /// </summary>
+        public Paddle? Winner7 { get; private set; }
+
+        /// <summary>
+        /// Eigth place winner paddle that was pressed. Null if no winner
+        /// </summary>
+        public Paddle? Winner8 { get; private set; }
 
         /// <summary>
         /// Reaction time for red 1 paddle in milliseconds. Null if no buzz in (paddle press) for red 1 paddle
@@ -55,7 +95,15 @@ namespace BriansUsbQuizBoxApi
         /// <summary>
         /// Constructor
         /// </summary>
-        /// <para name="winner">Winner paddle that was pressed. Null if no winner</para>
+        /// <param name="additionalWinnerInfo">True if additional winner info (beyond first place winner) is included, other false</param>
+        /// <param name="winner1">First place winner paddle that was pressed. Null if no winner</param>
+        /// <param name="winner2">Second winner paddle that was pressed. Null if no winner</param>
+        /// <param name="winner3">Third winner paddle that was pressed. Null if no winner</param>
+        /// <param name="winner4">Fourth winner paddle that was pressed. Null if no winner</param>
+        /// <param name="winner5">Fifth winner paddle that was pressed. Null if no winner</param>
+        /// <param name="winner6">Sixth winner paddle that was pressed. Null if no winner</param>
+        /// <param name="winner7">Seventh winner paddle that was pressed. Null if no winner</param>
+        /// <param name="winner8">Eight winner paddle that was pressed. Null if no winner</param>
         /// <param name="red1Time">Reaction time for red 1 paddle in milliseconds. Null if no buzz in (paddle press) for red 1 paddle</param>
         /// <param name="red2Time">Reaction time for red 2 paddle in milliseconds. Null if no buzz in (paddle press) for red 2 paddle</param>
         /// <param name="red3Time">Reaction time for red 3 paddle in milliseconds. Null if no buzz in (paddle press) for red 3 paddle</param>
@@ -64,11 +112,21 @@ namespace BriansUsbQuizBoxApi
         /// <param name="green2Time">Reaction time for green 2 paddle in milliseconds. Null if no buzz in (paddle press) for green 2 paddle</param>
         /// <param name="green3Time">Reaction time for green 3 paddle in milliseconds. Null if no buzz in (paddle press) for green 3 paddle</param>
         /// <param name="green4Time">Reaction time for green 4 paddle in milliseconds. Null if no buzz in (paddle press) for green 4 paddle</param>
-        public GameDoneEventArgs(Paddle? winner,
+        public GameDoneEventArgs(bool additionalWinnerInfo,
+            Paddle? winner1, Paddle? winner2, Paddle? winner3, Paddle? winner4,
+            Paddle? winner5, Paddle? winner6, Paddle? winner7, Paddle? winner8,
             decimal? red1Time, decimal? red2Time, decimal? red3Time, decimal? red4Time,
             decimal? green1Time, decimal? green2Time, decimal? green3Time, decimal? green4Time)
         {
-            Winner = winner;
+            HasAdditionalWinnerInfo = additionalWinnerInfo;
+            Winner = winner1;
+            Winner2 = winner2;
+            Winner3 = winner3;
+            Winner4 = winner4;
+            Winner5 = winner5;
+            Winner6 = winner6;
+            Winner7 = winner7;
+            Winner8 = winner8;
             Red1Time = red1Time;
             Red2Time = red2Time;
             Red3Time = red3Time;