+++ BriansUsbQuizBoxApi/Protocols/BoxStatusReport.cs
@@ -28,6 +28,50 @@ namespace BriansUsbQuizBoxApi.Protocols
 
         public decimal? Green4Time { get; private set; }
 
+        #region Protocol version 1 additions (document version 1.2)
+
+        /// <summary>
+        /// The communication protocol version
+        /// </summary>
+        public byte ProtocolVersion { get; private set; }
+
+        /// <summary>
+        /// Second place winner, will always be zero for protocol version 0
+        /// </summary>
+        public WinnerByte Winner2 { get; private set; }
+
+        /// <summary>
+        /// Third place winner, will always be zero for protocol version 0
+        /// </summary>
+        public WinnerByte Winner3 { get; private set; }
+
+        /// <summary>
+        /// Fourth place winner, will always be zero for protocol version 0
+        /// </summary>
+        public WinnerByte Winner4 { get; private set; }
+
+        /// <summary>
+        /// Fifth place winner, will always be zero for protocol version 0
+        /// </summary>
+        public WinnerByte Winner5 { get; private set; }
+
+        /// <summary>
+        /// Sixth place winner, will always be zero for protocol version 0
+        /// </summary>
+        public WinnerByte Winner6 { get; private set; }
+
+        /// <summary>
+        /// Seventh place winner, will always be zero for protocol version 0
+        /// </summary>
+        public WinnerByte Winner7 { get; private set; }
+
+        /// <summary>
+        /// Eigth place winner, will always be zero for protocol version 0
+        /// </summary>
+        public WinnerByte Winner8 { get; private set; }
+
+        #endregion
+
         /// <summary>
         /// Default constructor
         /// </summary>
@@ -36,7 +80,7 @@ namespace BriansUsbQuizBoxApi.Protocols
         }
 
         /// <summary>
-        /// Constructor for testing purposes only
+        /// Constructor for testing purposes only, protocol version 0
         /// </summary>
         /// <param name="status">Status byte</param>
         /// <param name="winner">Winner byte</param>
@@ -54,6 +98,51 @@ namespace BriansUsbQuizBoxApi.Protocols
         {
             Status = status;
             Winner = winner;
+            Red1Time = red1Time;
+            Red2Time = red2Time;
+            Red3Time = red3Time;
+            Red4Time = red4Time;
+            Green1Time = green1Time;
+            Green2Time = green2Time;
+            Green3Time = green3Time;
+            Green4Time = green4Time;
+        }
+
+        /// <summary>
+        /// Constructor for testing purposes only, protocol version 1
+        /// </summary>
+        /// <param name="status">Status byte</param>
+        /// <param name="winner1">First place winner byte</param>
+        /// <param name="winner2">Second place winner byte</param>
+        /// <param name="winner3">Third place winner byte</param>
+        /// <param name="winner4">Fourth place winner byte</param>
+        /// <param name="winner5">Fifth place winner byte</param>
+        /// <param name="winner6">Sixth place winner byte</param>
+        /// <param name="winner7">Seventh place winner byte</param>
+        /// <param name="winner8">Eight place winner byte</param>
+        /// <param name="red1Time">Time for red 1 paddle or null for no buzz in (paddle press)</param>
+        /// <param name="red2Time">Time for red 2 paddle or null for no buzz in (paddle press)</param>
+        /// <param name="red3Time">Time for red 3 paddle or null for no buzz in (paddle press)</param>
+        /// <param name="red4Time">Time for red 4 paddle or null for no buzz in (paddle press)</param>
+        /// <param name="green1Time">Time for green 1 paddle or null for no buzz in (paddle press)</param>
+        /// <param name="green2Time">Time for green 2 paddle or null for no buzz in (paddle press)</param>
+        /// <param name="green3Time">Time for green 3 paddle or null for no buzz in (paddle press)</param>
+        /// <param name="green4Time">Time for green 4 paddle or null for no buzz in (paddle press)</param>
+        public BoxStatusReport(StatusByte status,
+            WinnerByte winner1, WinnerByte winner2, WinnerByte winner3, WinnerByte winner4,
+            WinnerByte winner5, WinnerByte winner6, WinnerByte winner7, WinnerByte winner8,
+            decimal? red1Time, decimal? red2Time, decimal? red3Time, decimal? red4Time,
+            decimal? green1Time, decimal? green2Time, decimal? green3Time, decimal? green4Time)
+        {
+            Status = status;
+            Winner = winner1;
+            Winner2 = winner2;
+            Winner3 = winner3;
+            Winner4 = winner4;
+            Winner5 = winner5;
+            Winner6 = winner6;
+            Winner7 = winner7;
+            Winner8 = winner8;
             Red1Time = red1Time;
             Red2Time = red2Time;
             Red3Time = red3Time;
@@ -88,8 +177,16 @@ namespace BriansUsbQuizBoxApi.Protocols
                 return true;
             }
 
-            return Status == p.Status &&
+            return ProtocolVersion == p.ProtocolVersion &&
+                Status == p.Status &&
                 Winner == p.Winner &&
+                Winner2 == p.Winner2 &&
+                Winner3 == p.Winner3 &&
+                Winner4 == p.Winner4 &&
+                Winner5 == p.Winner5 &&
+                Winner6 == p.Winner6 &&
+                Winner7 == p.Winner7 &&
+                Winner8 == p.Winner8 &&
                 Red1Time == p.Red1Time &&
                 Red2Time == p.Red2Time &&
                 Red3Time == p.Red3Time &&
@@ -104,7 +201,9 @@ namespace BriansUsbQuizBoxApi.Protocols
         /// Generate hash code
         /// </summary>
         /// <returns>Hash code</returns>
-        public override int GetHashCode() => (Status, Winner, Red1Time, Red2Time, Red3Time, Red4Time,
+        public override int GetHashCode() => (ProtocolVersion, Status,
+            Winner, Winner2, Winner3, Winner4, Winner5, Winner6, Winner7, Winner8,
+            Red1Time, Red2Time, Red3Time, Red4Time,
             Green1Time, Green2Time, Green3Time, Green4Time).GetHashCode();
 
         /// <summary>
@@ -153,6 +252,7 @@ namespace BriansUsbQuizBoxApi.Protocols
                 throw new ArgumentOutOfRangeException(nameof(data), "Second byte expected to be 0x00");
             }
 
+            var protocolVersion = data[20];
             var retVal = new BoxStatusReport
             {
                 Status = (StatusByte)data[2],
@@ -164,17 +264,41 @@ namespace BriansUsbQuizBoxApi.Protocols
                 Green1Time = new byte[] { data[12], data[13] }.CalculateResponseTime(),
                 Green2Time = new byte[] { data[14], data[15] }.CalculateResponseTime(),
                 Green3Time = new byte[] { data[16], data[17] }.CalculateResponseTime(),
-                Green4Time = new byte[] { data[18], data[19] }.CalculateResponseTime()
+                Green4Time = new byte[] { data[18], data[19] }.CalculateResponseTime(),
+
+                #region Protocol version 1 additions (document version 1.2)
+
+                ProtocolVersion = protocolVersion,
+
+                // Need to force NO_VALID_WINNER (0x00) for protocol version 0 as Brian's Quiz Box fills non-zero values
+                Winner2 = protocolVersion == 0 ? WinnerByte.NO_VALID_WINNER : (WinnerByte)data[21],
+                Winner3 = protocolVersion == 0 ? WinnerByte.NO_VALID_WINNER : (WinnerByte)data[22],
+                Winner4 = protocolVersion == 0 ? WinnerByte.NO_VALID_WINNER : (WinnerByte)data[23],
+                Winner5 = protocolVersion == 0 ? WinnerByte.NO_VALID_WINNER : (WinnerByte)data[24],
+                Winner6 = protocolVersion == 0 ? WinnerByte.NO_VALID_WINNER : (WinnerByte)data[25],
+                Winner7 = protocolVersion == 0 ? WinnerByte.NO_VALID_WINNER : (WinnerByte)data[26],
+                Winner8 = protocolVersion == 0 ? WinnerByte.NO_VALID_WINNER : (WinnerByte)data[27],
+                
+                #endregion
             };
 
-            // Bytes 20-64 are don't cares...
+            // Bytes 28-64 are don't cares as of protocol version 1
 
             return retVal;
         }
 
         public override string ToString()
         {
-            return $"Status: {Status} (0x{Status:X}); Winner: {Winner} (0x{Winner:X}); "
+            return $"Protocol Version: {ProtocolVersion}; "
+                + $"Status: {Status} (0x{Status:X}); "
+                + $"Winner 1st: {Winner} (0x{Winner:X}); "
+                + $"Winner 2nd: {Winner2} (0x{Winner2:X}); "
+                + $"Winner 3rd: {Winner3} (0x{Winner3:X}); "
+                + $"Winner 4th: {Winner4} (0x{Winner4:X}); "
+                + $"Winner 5th: {Winner5} (0x{Winner5:X}); "
+                + $"Winner 6th: {Winner6} (0x{Winner6:X}); "
+                + $"Winner 7th: {Winner7} (0x{Winner7:X}); "
+                + $"Winner 8th: {Winner8} (0x{Winner8:X}); "
                 + $"Red 1 Time: {(Red1Time.HasValue ? Red1Time.Value.ToString() + "ms" : "[None]")}; "
                 + $"Red 2 Time: {(Red2Time.HasValue ? Red2Time.Value.ToString() + "ms" : "[None]")}; "
                 + $"Red 3 Time: {(Red3Time.HasValue ? Red3Time.Value.ToString() + "ms" : "[None]")}; "